requirements.txt
fastapi
uvicorn 
sqlalchemy 
psycopg2 
pydantic 
dotenv
alembic
passlib
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart